Add least effort models
These models can be accessed through the class AnomalyModels. Note that the least effort model uses gurobi, which is free for educational purposes: https://www.gurobi.com/ (note that you need to be connected to your university's network when activating the license). An example of how to use the model is provided.

Additionally, sketches of two more models are included -- targeted attack and random matrix -- though these are works in progress.