Fix various issues, use pyomo
The following fixes were applied (co-authored by Theodoros):
- The H matrix is now stored in a sparse format (for more efficient computations)
- State variables are now stored in radians instead of degrees
- State estimation now uses a W matrix as described in the literature
- Normalization of residual vectors has been fixed
- The anomaly detector now works in the presence of residual vectors containing inf values (which are treated as not anomalous)

Further, pyomo syntax is now used instead of gurobi. In addition to making the code a bit nicer, it should now be easier to switch solvers if this is desired in the future

Finally, PowerGrid.__init__() now supports all the test grids mentioned in the report outline (Illinois 200, PEGASE 1354, RTE 2848).